modules/location-producer/requirements.txt
grpcio==1.43.0
grpcio-tools==1.43.0
kafka-python==2.0.2
